Added context folder with UserContext file including logic. Also updated app.jsx logic.
