--- client/src/app.js
@@ -1,52 +1,0 @@
-import React, { useState } from 'react';
-import Navigation from './components/Navigation/Navigation';
-import NewDreamEntry from './components/NewDreamEntry';
-import PreviousDreams from './components/PreviousDreams';
-import DreamMeanings from './components/DreamMeanings';
-import './App.css';
-
-const App = () => {
-  const [user, setUser] = useState(null); // Placeholder for user authentication state
-  const [dreams, setDreams] = useState([]); // State for storing dream entries
-  const [isAuthenticated, setIsAuthenticated] = useState(false); // State to track if user is authenticated
-
-  // Function to handle user login
-  // Placeholder for actual login logic
-  const handleLogin = (username, password) => {
-    // Perform login operation...
-    setIsAuthenticated(true);
-    setUser({ username }); // Dummy user object
-  };
-
-  // Function to handle user logout
-  const handleLogout = () => {
-    // Perform logout operation...
-    setIsAuthenticated(false);
-    setUser(null);
-  };
-
-  // Function to save a new dream journal entry
-  const saveDreamEntry = (newDream) => {
-    setDreams([...dreams, newDream]);
-  };
-
-  return (
-    <div className="App">
-      <Navigation onLogout={handleLogout} isAuthenticated={isAuthenticated} />
-      {isAuthenticated ? (
-        <>
-          <NewDreamEntry onSaveDream={saveDreamEntry} />
-          <PreviousDreams dreams={dreams} />
-          <DreamMeanings />
-        </>
-      ) : (
-        // Placeholder for the Login component, to be replaced with actual login form
-        <div className="login-container">
-          <button onClick={() => handleLogin('user', 'pass')}>Login</button>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default App;
